Transform distance metrics (normalisation and log-scaling)

Internal helper function that applies optional transformations to distance metrics (mean and/or median) produced by [generate_geo_metrics_df()]. It can generate normalised and/or log-transformed versions of distance columns, depending on the options specified in `metric_type`. Normalisation parameters can be adjusted using `dnorm_args` and `dnorm_from` to control the shape and bounds of the weighting curve.

transform_distances(
  dist_df,
  metric_type,
  dnorm_args = c(mean = 0, sd = 50),
  dnorm_from = c(min = 0, max = 1e+05),
  round_val = 2
)

Arguments

dist_df

Data frame containing at least one of the columns `geo_min_dist_mean` and/or `geo_min_dist_median`, as generated by [generate_geo_metrics_df()].

metric_type

Character vector specifying which distance metrics and transformations to apply. Supports `"mean"`, `"median"`, `"norm"`, and `"log"`. - `"norm"` applies a Gaussian weighting (`dnorm`, mean = 0, SD = 50) so that nearby points have higher scores and distant ones much lower, then rescales using fixed `dnorm(0)` and `dnorm(1e5)` limits for comparability. - `"log"` applies a log-transform (with a 0.1 offset) and rescales to a 0 – 1 range.

dnorm_args

Named numeric vector giving the `mean` and `sd` parameters for the Gaussian weighting used in normalisation. Defaults to `c(mean = 0, sd = 50)`.

dnorm_from

Numeric vector of two values giving the range (in km) over which to anchor the scaling of the normalised scores. Defaults to `c(0, 1e5)`.

round_val

Numeric. Integer indicating the number of decimal places used for rounding results. Detaults to `2`.

Value

A data frame identical to `dist_df` with additional columns corresponding to the transformations requested: - `geo_min_dist_mean_norm` if `"mean"` and `"norm"` are included. - `geo_min_dist_median_norm` if `"median"` `"norm"` are included. - `geo_min_dist_mean_log` if `"mean"` and `"log"` are included. - `geo_min_dist_median_log` if `"median"` and `"log"` are included.